The 'freerdp' package should be a runtime dependency only.
The source configuration process of KRDC searches for the xfreerdp
binary and disables RDP support if the binary is not found.
I do not call that a "run-time dependency". But since xfreerdp actually
*is* a run-time dependency only, the SlackBuild will create a workaround.
With this modification you can add the 'freerdp' package afterwards and
RDP support will work out of the box without recompilation of 'krdc'.