Try to make kwin_wayland work on the console.
But I already realize that this is not the way (thanks to Martin Graesslin).
When using an xinitrc script, X.Org has already started when kwin_wayland
instantiates and therefore kwain will run as a nested window in the existing
X server.
Next attempt (also suggested by Martin Graesslin) is to run:
  export $(dbus-launch)
  startplasmacompositor
to start only wayland, without X.