Add applications-extra:yakuake
This commit is a bit of education on how to add a package on your own.
I don't do this with the intention of adding yakuake to my repository, but
to make it easier for people who have a need of it, to build it themselves.
Just download the latest stable tarball from:
  https://download.kde.org/stable/yakuake/
and move the tarball into ./kde/src/applications-extra/
Then run:
  cd kde
  ./KDE.SlackBuild applications-extra:yakuake
The resulting package can be found in /tmp/kde_build afterwards.