Add wayland support (not enabled in regular repo)
In an attempt to keep a single source tree for both 'latest' and 'testing'
repositories, everything to support a Plasma5 Wayland session was added.

This was made possible *without* the need for PAM or systemd because the
ConsoleKit2 and KWin developers cooperated (on my request :-) to add the
logind DBus API to CK2 and make KWin accept CK2 as an alternative to
systemd-logind if the latter is not present.

Note that this requires recompilation of two stock Slackware
packages, mesa and xorg-server, to let them gain Wayland support.

Build order:

Deps:
- ConsoleKit2
- wayland and wayland-protocols
- mesa
- xorg-server
- libxkbcommon
- qt5

Frameworks:
- kwayland
- plasma-framework

Plasma;
- kinfocenter
- kscreenlocker
- kwayland-integration
- libkscreen2
- plasma-desktop
- plasma-integration
- plasma-workspace
- powerdevil
- kwin

In order to start a Wayland session (you'll need to use packages from the
'testing' repository) either run 'startkwayland' from your console prompt in
runlevel 3, or choose the "Plasma (Wayland)" session in the SDDM dropdown
if you are booting into runlevel 4.