Disable applications/kpat package because it requires fc-solver
And fc-solver is not part of the Slackware distro. It in turn requires
additional packages and I am not prepared to add four new packages
to Slackware just to provide a Freecell solver for a Patience game.
Still wondering why the developer felt the need to impose a hard
dependency instead of making fc-solver an optional dep...