Updates to the dependencies for upcoming Plasma5 release
The update of icu4c in -current required that several packages had to be
recompiled or upgraded to get them to work again.
Unfortunately the update of icu4c in -current co-incided with my release
of KDE 5_18.10 and these deps had to be fixed post-release.

Some issues emerged as a result of these updates:
- SIP now stores private copies of sip.so in PyQt5 subdirectories,
  and the hp-systray in the hplip package fails to start as a result
- In the new QScintilla packages, I was unable to compile the
  Python2 bindings for PyQt4.