ktorrent: embed a local copy of the GeoIP database
The online version of GeoIP has been discontinued in favor of a new API.
The ktorrent code has not been updated for the new API and therefore
the program spits out a download error for geoip.dat everytime it starts.
By embedding a local copy of the database, a download is not attempted
and the error is gone.