Plasma-extra: added latte-dock
You can use Latte Dock as an alternative to the default Task Manager widget
in the Plasma5 panel.
To achieve this, you first have to move the default panel to the top of
the screen:
- Unlock widgets if needed.
- Right-click the panel and select "Panel Options > Configure Panel".
- Left-click and hold the "Screen Edge" button and drag the complete panel
  to the top of your screen.
Then remove the default Task Manager:
- Add a "spacer" widget next to the Task Manager.
- Hoever over the Task Manager area with your mouse and select "Remove".
  The spacer you just added will fill the vacated area.
Start Latte from the Plsama menu. It will add itself as an 'Autostart"
program so it will always be there when you login.
- Optionally configure Latte by starting programs and then pinning them to
  the dock, like with the default Task Manager.