Deps: add phonon support for kdelibs and Qt4 based applications
Used the last versions of phonon and phonon-gstreamer that still had
Qt4 support and renamed the packages to 'phonon-qt4' and
'phonon-qt4-gstreamer' so as not to make them clash with the Qt5 based
phonon packages.
Note that at leasr one backend is required to produce sound at all,
so I chose the gstreamer based backend because gstreamer is part of
Slackware and VLC is not.