Deps: add exiv2 and LibRaw updates to the official packages
Newer exiv2 is required to compile the latest Digikam 6.4.0.
Pat threw LibRaw in the mix for good measure.
Then he added the shared libraries of the official Slackware versions
of these packages to aaa_elflibs so that no breakage to official packages
will occur when you upgrade to their updated ktown versions.