Make kwallet-pam interact correctly with PAM
Using 'socat', kwallet-pam will now open the default wallet named 'kdewallet'
upon login, as long as its password is the same as the user's logon password.

For this to work, you must configure the Wallet Manager to never close the
open wallets (for instance after an application stops using the wallet),
because kwallet-pam is a one-shot action at login only.